refactor(compression): move clip context implementation into its own header
